change term from 'harga toko' to 'harga modal'

+++ halaman/tampil/barang_per_jenis_barang.php
@@ -5,7 +5,9 @@
         <h1 class="h3 mb-0 text-gray-800">Data Barang Jenis <?= $jenis_barang['nama']; ?></h1>
         <div>
             <a href="?h=barang" class="btn btn-secondary btn-sm"><i class="fas fa-caret-left"></i>&nbsp;&nbsp;Kembali</a>
-            <a href="?h=tambah_barang&id_jenis_barang=<?= $_GET['id_jenis_barang']; ?>" class="btn btn-primary btn-sm"><i class="fas fa-plus"></i>&nbsp;&nbsp;Tambah</a>
+            <?php if ($_SESSION['user']['status'] == 'ADMIN') : ?>
+                <a href="?h=tambah_barang&id_jenis_barang=<?= $_GET['id_jenis_barang']; ?>" class="btn btn-primary btn-sm"><i class="fas fa-plus"></i>&nbsp;&nbsp;Tambah</a>
+            <?php endif; ?>
         </div>
     </div>
 
@@ -31,10 +33,12 @@
                                     <th class="td-fit text-center">No</th>
                                     <th class="text-center">Kode Barang</th>
                                     <th class="text-center">Nama Barang</th>
-                                    <th class="text-center">Harga Toko</th>
+                                    <th class="text-center">Harga Modal</th>
                                     <th class="text-center">Harga Label</th>
                                     <th class="text-center">Stok</th>
-                                    <th class="td-fit text-center">Aksi</th>
+                                    <?php if ($_SESSION['user']['status'] == 'ADMIN') : ?>
+                                        <th class="td-fit text-center">Aksi</th>
+                                    <?php endif; ?>
                                 </tr>
                             </thead>
                             <?php
@@ -45,15 +49,17 @@
                                 <?php while ($row = $result->fetch_assoc()) : ?>
                                     <tr>
                                         <td class="td-fit align-middle text-center"><?= $no++; ?></td>
-                                        <td class="align-middle text-center"><?= $row['kode_jenis_barang'].generateKodeBarang($row['kode']); ?></td>
+                                        <td class="align-middle text-center"><?= $row['kode_jenis_barang'] . generateKodeBarang($row['kode']); ?></td>
                                         <td class="align-middle text-center"><?= $row['nama']; ?></td>
                                         <td class="align-middle text-center"><?= number_format($row['harga_toko'], 0, ",", "."); ?></td>
                                         <td class="align-middle text-center"><?= number_format($row['harga_label'], 0, ",", "."); ?></td>
                                         <td class="align-middle text-center"><?= $row['stok']; ?> <?= $row['satuan']; ?></td>
-                                        <td class="text-center td-fit">
-                                            <a href="?h=edit_barang&id=<?= $row['id']; ?>&id_jenis_barang=<?= $row['id_jenis_barang']; ?>" class="btn btn-sm btn-warning"><i class="far fa-edit"></i></a>
-                                            <a href="?h=hapus_barang&id=<?= $row['id']; ?>&id_jenis_barang=<?= $row['id_jenis_barang']; ?>" class="btn btn-sm btn-danger" onclick="return confirm('Yakin ingin menghapus data ini?')"><i class="far fa-trash-alt"></i></a>
-                                        </td>
+                                        <?php if ($_SESSION['user']['status'] == 'ADMIN') : ?>
+                                            <td class="text-center td-fit">
+                                                <a href="?h=edit_barang&id=<?= $row['id']; ?>&id_jenis_barang=<?= $row['id_jenis_barang']; ?>" class="btn btn-sm btn-warning"><i class="far fa-edit"></i></a>
+                                                <a href="?h=hapus_barang&id=<?= $row['id']; ?>&id_jenis_barang=<?= $row['id_jenis_barang']; ?>" class="btn btn-sm btn-danger" onclick="return confirm('Yakin ingin menghapus data ini?')"><i class="far fa-trash-alt"></i></a>
+                                            </td>
+                                        <?php endif; ?>
                                     </tr>
                                 <?php endwhile; ?>
                             </tbody>
